feat: Add Swift Package Manager (SPM) support
Migrate iOS native sources from Classes/ to the SPM-standard layout
(ios/flutter_credit_card/Sources/flutter_credit_card/) so Flutter 3.24+
can resolve this plugin via Swift Package Manager without CocoaPods.

- Add ios/flutter_credit_card/Package.swift with swift-tools-version 5.9
- Move all Swift sources to ios/flutter_credit_card/Sources/flutter_credit_card/
- Update podspec source_files path to the new location; drop deprecated
  i386 simulator exclusion; bump min iOS from 11.0 → 12.0 (matches SPM target)
- CocoaPods path continues to work unchanged for apps not yet on SPM

Closes #196
