--- ios/Classes/Constants.swift
@@ -1,7 +1,0 @@
-enum Constants {
-    static let gyroEventChannel = "com.simform.flutter_credit_card/gyroscope"
-    static let gyroMethodChannel = "com.simform.flutter_credit_card"
-    static let initiateEvents = "initiateEvents"
-    static let cancelEvents = "cancelEvents"
-    static let isGyroAvailable = "isGyroscopeAvailable"
-}
