+++ ios/flutter_credit_card.podspec
@@ -14,11 +14,10 @@ and float.
   s.license          = { :type => 'MIT', :file => '../LICENSE' }
   s.author           = { 'Simform Solutions' => 'developer@simform.com' }
   s.source           = { :http => 'https://github.com/SimformSolutionsPvtLtd/flutter_credit_card/tree/master' }
-  s.source_files = 'Classes/**/*'
+  s.source_files = 'flutter_credit_card/Sources/flutter_credit_card/**/*.swift'
   s.dependency 'Flutter'
-  s.platform = :ios, '11.0'
+  s.platform = :ios, '12.0'
 
-  # Flutter.framework does not contain a i386 slice.
-  s.pod_target_xcconfig = { 'DEFINES_MODULE' => 'YES', 'EXCLUDED_ARCHS[sdk=iphonesimulator*]' => 'i386' }
+  s.pod_target_xcconfig = { 'DEFINES_MODULE' => 'YES' }
   s.swift_version = '5.0'
 end
