+++ ios/flutter_credit_card/Sources/flutter_credit_card/Constants.swift
@@ -1,0 +1,7 @@
+enum Constants {
+  static let gyroEventChannel = "com.simform.flutter_credit_card/gyroscope"
+  static let gyroMethodChannel = "com.simform.flutter_credit_card"
+  static let initiateEvents = "initiateEvents"
+  static let cancelEvents = "cancelEvents"
+  static let isGyroAvailable = "isGyroscopeAvailable"
+}
