+++ ios/flutter_credit_card/Sources/flutter_credit_card/FlutterCreditCardPlugin.swift
@@ -1,0 +1,15 @@
+import CoreMotion
+import Flutter
+
+public class FlutterCreditCardPlugin: NSObject, FlutterPlugin {
+  private var gyroscopeChannel: GyroscopeChannelImpl?
+
+  public static func register(with registrar: FlutterPluginRegistrar) {
+    let instance = FlutterCreditCardPlugin()
+    instance.gyroscopeChannel = GyroscopeChannelImpl(messenger: registrar.messenger())
+  }
+
+  public func detachFromEngine(for registrar: FlutterPluginRegistrar) {
+    gyroscopeChannel = nil
+  }
+}
